+++ gulpfile.babel.js
@@ -1,27 +1,46 @@
 import gulp from "gulp";
+import gutil from "gulp-util";
 import browserify from "browserify";
+import rename from 'gulp-rename';
+import flatten from 'gulp-flatten';
+import es from 'event-stream';
 import source from "vinyl-source-stream";
 import babelify from "babelify";
 import vueify from "vueify";
 import uglify from "gulp-uglify";
 import watchify from "watchify";
 import buffer from "vinyl-buffer";
+import {argv} from 'yargs';
 import { Server } from "karma";
 
-const entry = "src/dist.js";
+console.log(argv.f);
+
+const entry = (argv.f !==undefined) ? argv.f : "./src/reg/star-rating.js";
 const dest = "dist"; // destination folder
 
-gulp.task('default', () => {
-    browserify({
-            entries: entry,
-            debug: true
-        })
-        .bundle()
-        .pipe(source('star-rating.min.js'))
-        .pipe(buffer())
-        .pipe(uglify())
-        .pipe(gulp.dest(dest));
+gulp.task('default', function(done) {
+    gulp.src(['./src/reg/**-rating.js'], function(err, files) {
+        if (err) done(err);
+
+        var tasks = files.map(function(entry) {
+            return browserify({
+                    entries: [entry]
+                })
+                .external('vue') // exclude vue from dist files, this will be shimmed globally
+                .bundle()
+                .pipe(source(entry))
+                .pipe(flatten())
+                .pipe(rename({
+                    extname: '.min.js'
+                }))
+                .pipe(buffer())
+                .pipe(uglify())
+                .pipe(gulp.dest(dest));
+        });
+        es.merge(tasks).on('end', done);
+    })
 });
+
 
 gulp.task('watch', () => {
     const b = browserify({
@@ -30,9 +49,11 @@ gulp.task('watch', () => {
         debug: true,
         cache: {},
         packageCache: {}
-    })
+    });
 
+    b.external('vue');
     b.on('update', bundle);
+
     bundle();
 
     function bundle() {
@@ -51,6 +72,6 @@ gulp.task('watch', () => {
 gulp.task('test', done => {
     new Server({
         configFile: __dirname + '/karma.conf.js',
-        singleRun: true
+        singleRun: false
     }, done).start();
 });