+++ src/raters/stars/star-rating.vue
@@ -1,6 +1,7 @@
 <script type="text/javascript">
-import BaseRating from './raters/BaseRating.vue'
-import Star from './Star.vue';
+import BaseRating from '../BaseRating.vue';
+import Star from './star.vue';
+
 export default BaseRating.extend({
     name: 'Star-Rating',
     components: {